feat: add clerky as an authentication option
Closes #1336

chore: make it possible to switch between clerky and kratos

Closes #1326

feat: add sign and sign up forms from clerky

feat: add wizard guide for the sign up process

fix: fix failing build

fix: provide user for clerky and kratos seperately

fix: add wrapper for session checker for clerky

feat: ensure logout button works with clerky

fix: add authorization header for clerky token

fix: remove bearer token attachment

fix: issue with incorrect redirect

fix: use react router or next router instead of window location

fix: fix issue with redirecting to login page

chore: rename from clerky to clerk
