+++ .env
@@ -5,3 +5,4 @@ PORT=3004
 # NEXT_PUBLIC_WITHOUT_SESSION=true
 # BACKEND_URL=https://incident-commander.demo.aws.flanksource.com/
 # ENV=
+# NEXT_PUBLIC_AUTH_IS_CLERK=true
