+++ pages/error.tsx
@@ -1,62 +1,13 @@
-import { SelfServiceError } from "@ory/client";
-import { AxiosError } from "axios";
-import type { NextPage } from "next";
-import Link from "next/link";
-import { useRouter } from "next/router";
-import { useEffect, useState } from "react";
+import KratosErrorPage from "../src/components/Authentication/Kratos/KratosError";
+import useDetermineAuthSystem from "../src/components/Authentication/useDetermineAuthSystem";
 
-import ory from "../src/components/ory/sdk";
+export default function ErrorPage() {
+  const authSystem = useDetermineAuthSystem();
 
-const Login: NextPage = () => {
-  const [error, setError] = useState<SelfServiceError | string>();
-
-  // Get ?id=... from the URL
-  const router = useRouter();
-  const { id } = router.query;
-
-  useEffect(() => {
-    // If the router is not ready yet, or we already have an error, do nothing.
-    if (!router.isReady || error) {
-      return;
-    }
-
-    ory
-      .getSelfServiceError(String(id))
-      .then(({ data }) => {
-        setError(data);
-      })
-      .catch((err: AxiosError) => {
-        switch (err.response?.status) {
-          case 404:
-          // The error id could not be found. Let's just redirect home!
-          case 403:
-          // The error id could not be fetched due to e.g. a CSRF issue. Let's just redirect home!
-          case 410:
-            // The error id expired. Let's just redirect home!
-            return router.push("/");
-        }
-
-        return Promise.reject(err);
-      });
-  }, [id, router, router.isReady, error]);
-
-  if (!error) {
+  if (authSystem === "clerk") {
+    // TODO: Clerk login
     return null;
   }
 
-  return (
-    <>
-      <div>
-        <div>An error occurred</div>
-        <div>{JSON.stringify(error, null, 2)}</div>
-      </div>
-      <div>
-        <Link href="/" passHref>
-          <div>Go back</div>
-        </Link>
-      </div>
-    </>
-  );
-};
-
-export default Login;
+  return <KratosErrorPage />;
+}
