+++ src/components/Authentication/AuthSessionChecker.tsx
@@ -1,0 +1,19 @@
+import useDetermineAuthSystem from "./useDetermineAuthSystem";
+import ClerkAuthSessionChecker from "./Clerk/ClerkAuthSessionChecker";
+import KratosAuthSessionChecker from "./Kratos/KratosAuthSessionChecker";
+
+type AuthSessionCheckerProps = {
+  children: React.ReactNode;
+};
+
+export default function AuthSessionChecker({
+  children
+}: AuthSessionCheckerProps) {
+  const authSystem = useDetermineAuthSystem();
+
+  if (authSystem === "clerk") {
+    return <ClerkAuthSessionChecker>{children}</ClerkAuthSessionChecker>;
+  }
+
+  return <KratosAuthSessionChecker>{children}</KratosAuthSessionChecker>;
+}
