+++ src/components/Authentication/Kratos/KratosVerification.tsx
@@ -1,0 +1,114 @@
+import {
+  SelfServiceVerificationFlow,
+  SubmitSelfServiceVerificationFlowBody
+} from "@ory/client";
+import { AxiosError } from "axios";
+import type { NextPage } from "next";
+import Head from "next/head";
+import Link from "next/link";
+import { useRouter } from "next/router";
+import { useEffect, useState } from "react";
+import { Flow } from "../../ory";
+import ory from "../../ory/sdk";
+
+function KratosVerification() {
+  const [flow, setFlow] = useState<SelfServiceVerificationFlow>();
+
+  // Get ?flow=... from the URL
+  const router = useRouter();
+  const { flow: flowId, return_to: returnTo } = router.query;
+
+  useEffect(() => {
+    // If the router is not ready yet, or we already have a flow, do nothing.
+    if (!router.isReady || flow) {
+      return;
+    }
+
+    // If ?flow=.. was in the URL, we fetch it
+    if (flowId) {
+      ory
+        .getSelfServiceVerificationFlow(String(flowId))
+        .then(({ data }) => {
+          setFlow(data);
+        })
+        .catch((err: AxiosError) => {
+          switch (err.response?.status) {
+            case 410:
+            // Status code 410 means the request has expired - so let's load a fresh flow!
+            case 403:
+              // Status code 403 implies some other issue (e.g. CSRF) - let's reload!
+              return router.push("/verification");
+          }
+
+          throw err;
+        });
+      return;
+    }
+
+    // Otherwise we initialize it
+    ory
+      .initializeSelfServiceVerificationFlowForBrowsers(
+        returnTo ? String(returnTo) : undefined
+      )
+      .then(({ data }) => {
+        setFlow(data);
+      })
+      .catch((err: AxiosError) => {
+        switch (err.response?.status) {
+          case 400:
+            // Status code 400 implies the user is already signed in
+            return router.push("/");
+        }
+
+        throw err;
+      });
+  }, [flowId, router, router.isReady, returnTo, flow]);
+
+  const onSubmit = (values: SubmitSelfServiceVerificationFlowBody) =>
+    router
+      // On submission, add the flow ID to the URL but do not navigate. This prevents the user loosing
+      // his data when she/he reloads the page.
+      .push(`/verification?flow=${flow?.id}`, undefined, { shallow: true })
+      .then(() =>
+        ory
+          .submitSelfServiceVerificationFlow(
+            String(flow?.id),
+            undefined,
+            values
+          )
+          .then(({ data }) => {
+            // Form submission was successful, show the message to the user!
+            setFlow(data);
+          })
+          .catch((err: AxiosError) => {
+            switch (err.response?.status) {
+              case 400:
+                // Status code 400 implies the form validation had an error
+                setFlow(err.response?.data);
+                return;
+            }
+
+            throw err;
+          })
+      );
+
+  return (
+    <>
+      <Head>
+        <title>Verify your account - Ory NextJS Integration Example</title>
+        <meta name="description" content="NextJS + React + Vercel + Ory" />
+      </Head>
+      <div>
+        <div>Verify your account</div>
+        <Flow onSubmit={onSubmit} flow={flow} />
+      </div>
+      <div>
+        <Link href="/" passHref>
+          <div>Go back</div>
+        </Link>
+      </div>
+    </>
+  );
+}
+
+export default KratosVerification;
