+++ src/components/ory/ui/NodeAnchor.tsx
@@ -1,5 +1,6 @@
 import { UiNodeAnchorAttributes } from "@ory/client";
 import { UiNode } from "@ory/client";
+import { useRouter } from "next/router";
 
 interface Props {
   node: UiNode;
@@ -7,13 +8,15 @@ interface Props {
 }
 
 export const NodeAnchor = ({ attributes }: Props) => {
+  const { push } = useRouter();
+
   return (
     <button
       className="font-medium text-indigo-600 hover:text-indigo-500"
       onClick={(e) => {
         e.stopPropagation();
         e.preventDefault();
-        window.location.href = attributes.href;
+        push(attributes.href);
       }}
     >
       {attributes.title.text}
